메인카테고리별분류 로직적용
