+++ src/components/categories/Categories.tsx
@@ -18,64 +18,78 @@ export const categories = [
     path: "게임",
     icon: SiDuckduckgo,
     description: "게임프로젝트 입니다.",
+    value: 1,
   },
   {
     label: "여행 프로젝트",
     path: "여행",
     icon: SiTourbox,
     description: "여행 프로젝트입니다.",
+    value: 2,
   },
   {
     label: "운동 프로젝트",
     path: "운동",
     icon: MdSportsHandball,
     description: "운동 프로젝트입니다.",
+    value: 3,
   },
   {
     label: "책 프로젝트",
     path: "책",
     icon: FaBookTanakh,
     description: "책 프로젝트입니다.",
+    value: 4,
   },
   {
     label: "직무 프로젝트",
     path: "직무",
     icon: GrWorkshop,
     description: "직무 프로젝트입니다.",
+    value: 5,
   },
   {
     label: "언어 프로젝트",
     path: "언어",
     icon: LiaChalkboardTeacherSolid,
     description: "언어 프로젝트입니다.",
+    value: 6,
   },
   {
     label: "공연 프로젝트",
     path: "공연",
     icon: SiThemoviedatabase,
     description: "공연 프로젝트입니다.",
+    value: 7,
   },
   {
     label: "음악 프로젝트",
     path: "음악",
     icon: SiYoutubemusic,
     description: "음악 프로젝트입니다.",
+    value: 8,
   },
   {
     label: "공예 프로젝트",
     path: "공예",
     icon: FaHands,
     description: "공예 프로젝트입니다.",
+    value: 9,
   },
   {
     label: "댄스 프로젝트",
     path: "댄스",
     icon: SiBytedance,
     description: "댄스 프로젝트입니다.",
+    value: 10,
   },
 ];
 
-const Categories = () => {
+interface CategoriesProps {
+  onSelect: (value: number) => void; // onSelect prop의 타입을 명시
+}
+
+const Categories: React.FC<CategoriesProps> = ({ onSelect }) => {
   const { category } = useParams();
 
   return (
@@ -87,6 +101,8 @@ const Categories = () => {
           path={item.path}
           icon={item.icon}
           selected={category === item.path}
+          value={item.value}
+          onSelect={onSelect}
         />
       ))}
     </StCategory>
